[DT-21] Change version code type from Int to Long
To match the actual android framework type for the same.

+++ install-manager/src/main/java/cm/aptoide/pt/install_manager/RealApp.kt
@@ -43,7 +43,7 @@ internal class RealApp<D> internal constructor(
         throw IllegalArgumentException("This version is already installed")
       }
       installPackageInfo.version.versionCode < (_installedVersion?.versionCode
-        ?: Int.MIN_VALUE) -> {
+        ?: Long.MIN_VALUE) -> {
         throw IllegalArgumentException("Newer version is installed")
       }
       else -> {
